requirements.txt
Flask==3.0.3
openai==1.3.5
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.0
openpyxl==3.1.2
gunicorn==21.2.0
nginx>=1.24.0
